requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
imagehash>=4.3.1
numpy>=1.24.0
tqdm>=4.65.0

# ── Opcionales para modo GPU (--gpu) ──
# Descomenta las siguientes líneas o instala manualmente:
# torch>=2.0.0
# torchvision>=0.15.0
